refactor: add `GenerateStage#render_chunk_to_assets`
