+++ crates/rolldown/src/stages/generate_stage/mod.rs
@@ -5,12 +5,7 @@ use oxc::{ast::VisitMut, index::IndexVec};
 use rolldown_ecmascript::AstSnippet;
 use rustc_hash::FxHashSet;
 
-use futures::future::try_join_all;
-use rolldown_common::{
-  ChunkIdx, ChunkKind, FileNameRenderOptions, Module, Output, OutputAsset, OutputChunk,
-  PreliminaryFilename, SourceMapType,
-};
-use rolldown_error::BuildError;
+use rolldown_common::{ChunkIdx, ChunkKind, FileNameRenderOptions, Module, PreliminaryFilename};
 use rolldown_plugin::SharedPluginDriver;
 use rolldown_utils::{
   path_buf_ext::PathBufExt,
@@ -27,23 +22,17 @@ use crate::{
   },
   stages::link_stage::LinkStageOutput,
   utils::{
-    augment_chunk_hash::augment_chunk_hash,
-    chunk::{
-      deconflict_chunk_symbols::deconflict_chunk_symbols,
-      finalize_chunks::finalize_chunks,
-      render_chunk::{render_chunk, ChunkRenderReturn},
-    },
+    chunk::deconflict_chunk_symbols::deconflict_chunk_symbols,
     extract_hash_pattern::extract_hash_pattern,
     extract_meaningful_input_name_from_path::try_extract_meaningful_input_name_from_path,
-    finalize_normal_module,
-    hash_placeholder::HashPlaceholderGenerator,
-    render_chunks::render_chunks,
+    finalize_normal_module, hash_placeholder::HashPlaceholderGenerator,
   },
   BundleOutput, SharedOptions,
 };
 
 mod code_splitting;
 mod compute_cross_chunk_links;
+mod render_chunk_to_assets;
 
 pub struct GenerateStage<'a> {
   link_output: &'a mut LinkStageOutput,
@@ -122,110 +111,7 @@ impl<'a> GenerateStage<'a> {
         }
       });
 
-    let chunks = try_join_all(
-      chunk_graph
-        .chunks
-        .iter()
-        .map(|c| async { render_chunk(c, self.options, self.link_output, &chunk_graph).await }),
-    )
-    .await?;
-
-    let chunks = render_chunks(self.plugin_driver, chunks).await?;
-
-    let chunks = augment_chunk_hash(self.plugin_driver, chunks).await?;
-
-    let chunks = finalize_chunks(&mut chunk_graph, chunks);
-
-    let mut assets = vec![];
-    for ChunkRenderReturn {
-      mut map,
-      rendered_chunk,
-      mut code,
-      file_dir,
-      preliminary_filename,
-      ..
-    } in chunks
-    {
-      if let Some(map) = map.as_mut() {
-        map.set_file(&rendered_chunk.filename);
-
-        let map_filename = format!("{}.map", rendered_chunk.filename.as_str());
-        let map_path = file_dir.join(&map_filename);
-
-        if let Some(source_map_ignore_list) = &self.options.sourcemap_ignore_list {
-          let mut x_google_ignore_list = vec![];
-          for (index, source) in map.get_sources().enumerate() {
-            if source_map_ignore_list.call(source, map_path.to_string_lossy().as_ref()).await? {
-              #[allow(clippy::cast_possible_truncation)]
-              x_google_ignore_list.push(index as u32);
-            }
-          }
-          if !x_google_ignore_list.is_empty() {
-            map.set_x_google_ignore_list(x_google_ignore_list);
-          }
-        }
-
-        if let Some(sourcemap_path_transform) = &self.options.sourcemap_path_transform {
-          let mut sources = Vec::with_capacity(map.get_sources().count());
-          for source in map.get_sources() {
-            sources.push(
-              sourcemap_path_transform.call(source, map_path.to_string_lossy().as_ref()).await?,
-            );
-          }
-          map.set_sources(sources.iter().map(std::convert::AsRef::as_ref).collect::<Vec<_>>());
-        }
-
-        // Normalize the windows path at final.
-        let sources = map.get_sources().map(|x| x.to_slash_lossy().to_string()).collect::<Vec<_>>();
-        map.set_sources(sources.iter().map(std::convert::AsRef::as_ref).collect::<Vec<_>>());
-
-        match self.options.sourcemap {
-          SourceMapType::File => {
-            let source = match map.to_json_string().map_err(BuildError::sourcemap_error) {
-              Ok(source) => source,
-              Err(e) => {
-                self.link_output.errors.push(e);
-                continue;
-              }
-            };
-            assets.push(Output::Asset(Box::new(OutputAsset {
-              filename: map_filename.clone(),
-              source: source.into(),
-            })));
-            code.push_str(&format!("\n//# sourceMappingURL={map_filename}"));
-          }
-          SourceMapType::Inline => {
-            let data_url = match map.to_data_url().map_err(BuildError::sourcemap_error) {
-              Ok(data_url) => data_url,
-              Err(e) => {
-                self.link_output.errors.push(e);
-                continue;
-              }
-            };
-            code.push_str(&format!("\n//# sourceMappingURL={data_url}"));
-          }
-          SourceMapType::Hidden => {}
-        }
-      }
-      let sourcemap_filename =
-        map.as_ref().map(|_| format!("{}.map", rendered_chunk.filename.as_str()));
-      assets.push(Output::Chunk(Box::new(OutputChunk {
-        name: rendered_chunk.name,
-        filename: rendered_chunk.filename,
-        code,
-        is_entry: rendered_chunk.is_entry,
-        is_dynamic_entry: rendered_chunk.is_dynamic_entry,
-        facade_module_id: rendered_chunk.facade_module_id,
-        modules: rendered_chunk.modules,
-        exports: rendered_chunk.exports,
-        module_ids: rendered_chunk.module_ids,
-        imports: rendered_chunk.imports,
-        dynamic_imports: rendered_chunk.dynamic_imports,
-        map,
-        sourcemap_filename,
-        preliminary_filename: preliminary_filename.to_string(),
-      })));
-    }
+    let mut assets = self.render_chunk_to_assets(&mut chunk_graph).await?;
 
     // Make sure order of assets are deterministic
     // TODO: use `preliminary_filename` on `Output::Asset` instead
